pyproject.toml
[project]
name = "writer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
prerelease = "allow"
dependencies = [
    "alembic>=1.16.5",
    "fastapi-cli>=0.0.11",
    "fastapi[all,standard]>=0.116.1",
    "httpx>=0.28.1",
    "ipython>=9.5.0",
    "jinja2>=3.1.6",
    "mypy>=1.18.2",
    "openai>=1.107.0",
    "orjson>=3.11.3",
    "pydantic-extra-types>=2.10.5",
    "pydantic-settings>=2.0.0",
    "pydantic>=2.12.0a1",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "ruff>=0.12.12",
    "sqlalchemy[asyncio]>=2.0.43",
    "sqlmodel>=0.0.25",
    "tzdata>=2025.2",
    "uvicorn[standard]>=0.35.0",
    "uvloop>=0.21.0",
    "websockets>=15.0.1",
    "pyyaml>=6.0.3",
    "writer-sdk>=2.3.2",
    "psycopg>=3.2.10",
]

[tool.uv]
prerelease = "allow"

[tool.mypy]
plugins = ["pydantic.mypy"]

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

